server/pyproject.toml
[project]
name = "src"
version = "0.1.0"
description = ""
authors = [
    {name = "GiovanniDonati",email = "grdonati99@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13, <4.0"
dependencies = [
    "fastapi[standard] (>=0.128.0,<0.129.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "sqlalchemy (>=2.0.46,<3.0.0)",
    "alembic (>=1.18.1,<2.0.0)",
    "pymysql (>=1.1.1)",
    "mysql (>=0.0.3)",
    "mysqlclient (>=2.2.6)",
    "cryptography (>=45.0.6)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "pwdlib[argon2] (>=0.3.0,<0.4.0)"
    ]

[tool.poetry.group.dev.dependencies]
pytest = "^9.0.2"
pytest-cov = "^7.0.0"
taskipy = "^1.14.1"

[tool.pytest.ini_options]
pythonpath = "."

[tool.taskipy.tasks]
run = 'set PYTHONPATH=src&& fastapi dev src/app.py'
test = 'set PYTHONPATH=src&& pytest -s -x --cov=src -vv'
post_test = 'coverage html'

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
